requirements.txt
# Framework Django
Django>=4.2,<5.0

# Pour l'environnement de développement
django-environ  # Pour la gestion des variables d'environnement
gunicorn        # Serveur WSGI (pour le déploiement en production)

# Outils complémentaires (si nécessaire)
djangorestframework  # Pour développer des APIs REST
Pillow               # Pour le traitement d'images
django-cors-headers
# Outils de tests
pytest-django
